refactor: Clean up test files
- Remove non-working tests (do or do not approach)
- Remove '_simple' naming convention
- Remove obvious/redundant comments
- Rename tests/README.md to tests/CLAUDE.md
- Fix missing newlines at end of files
